feat: configured kacy menu button
